# openapi.yaml
openapi: 3.1.0
info:
  title: Harvist — Scraper de Imágenes
  version: 1.0.0
  description: |
    API REST para la descarga masiva y automatizada de imágenes de productos
    a partir de un CSV de inventario. Busca imágenes en Bing Images mediante
    un patrón Productor/Consumidor, las valida, redimensiona y las sirve al
    usuario como archivo ZIP descargable.

    ## Autores
    - BenjaminDTS
    - Carlos Vico

    ## WebSocket
    El endpoint `GET /api/v1/jobs/{job_id}/ws` se gestiona mediante protocolo
    WebSocket (ws:// en desarrollo, wss:// en producción). OpenAPI 3.1 no
    tiene soporte nativo de WebSocket; consulta la sección `x-websocket` de
    cada operación para el contrato de mensajes.

    ## Respuesta estándar
    Todas las respuestas JSON siguen el contrato:
    ```json
    { "success": true, "data": {}, "message": "Descripción" }
    ```
  x-author: BenjaminDTS
  contact:
    name: BenjaminDTS
  license:
    name: Propietaria
servers:
  - url: http://localhost:8000
    description: Entorno de desarrollo local
  - url: https://staging.harvist.app
    description: Entorno de staging
  - url: https://harvist.app
    description: Entorno de producción
...
paths:
...
  /api/v1/files/{job_id}:
    get:
      tags:
        - files
      summary: Descargar el ZIP de imágenes de un trabajo completado
      description: |
        Devuelve el archivo ZIP generado para el trabajo indicado.
        Solo disponible cuando `estado` es `completado`.
        Si `generar_descripciones` estuvo activo, el ZIP incluye también
        `descripciones.csv` con los campos `codigo`, `nombre` y
        `descripcion_generada`.
      operationId: downloadJobZip
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Archivo ZIP del trabajo.
          headers:
            Content-Disposition:
              description: 'attachment; filename="{job_id}.zip"'
              schema:
                type: string
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          description: El ZIP aún no existe o el trabajo no está completado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: ZIP no disponible
                  value:
                    success: false
                    data: null
                    message: El archivo ZIP no está disponible todavía.

    delete:
      tags:
        - files
      summary: Eliminar los archivos generados de un trabajo
      description: |
        Elimina el directorio de imágenes y el archivo ZIP asociados al
        trabajo. Útil para liberar espacio en disco tras la descarga.
      operationId: deleteJobFiles
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Archivos eliminados correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDeleteResponse'
              examples:
                success:
                  summary: Eliminación correcta
                  value:
                    success: true
                    data:
                      job_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    message: Archivos eliminados correctamente.
        '404':
          description: No existen archivos para el trabajo indicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: Archivos no encontrados
                  value:
                    success: false
                    data: null
                    message: No existen archivos para el trabajo especificado.
        '500':
          description: Error interno al intentar eliminar los archivos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internal_error:
                  summary: Error de sistema de archivos
                  value:
                    success: false
                    data: null
                    message: No se pudieron eliminar los archivos. Consulte los logs del servidor.
...
components:
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      description: Identificador único del trabajo (UUID v4).
      schema:
        type: string
        format: uuid
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
  schemas:
...
    JobDeleteData:
      type: object
      description: Datos devueltos al eliminar archivos de un trabajo.
      required:
        - job_id
      properties:
        job_id:
          type: string
          format: uuid
          description: Identificador del trabajo cuyos archivos fueron eliminados.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6

    # ── Eventos WebSocket ──────────────────────────────────────────────────
...
    JobDeleteResponse:
      type: object
      description: Respuesta estándar para la eliminación de archivos de un trabajo.
      required:
        - success
        - data
        - message
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/JobDeleteData'
        message:
          type: string
          example: Archivos eliminados correctamente.
    ErrorResponse:
      type: object
      description: Respuesta estándar para todos los errores de la API.
      required:
        - success
        - data
        - message
      properties:
        success:
          type: boolean
          example: false
        data:
          type:
            - object
            - 'null'
          description: Siempre `null` en respuestas de error.
          example: null
        message:
          type: string
          description: Descripción legible del error. Nunca incluye stack traces en producción.
          example: El recurso solicitado no existe.
...